chore(headless): remove unnecessary dependencies
