--- packages/shared/src/index.ts
@@ -1,2 +1,0 @@
-// export * from './utils/use-logical-keys';
-export * from './utils/use-ordinal';
