--- packages/shared/src/utils/use-ordinal.ts
@@ -1,14 +1,0 @@
-import { noSerialize } from '@builder.io/qwik';
-
-export const useOrdinal = () => {
-  return noSerialize((n: number) => {
-    const pr = new Intl.PluralRules('en-GB', { type: 'ordinal' });
-    const suffixes = new Map([
-      ['one', 'st'],
-      ['two', 'nd'],
-      ['few', 'rd'],
-      ['other', 'th'],
-    ]);
-    return `${n}${suffixes.get(pr.select(n))}`;
-  });
-};
